[Zooming Container] Card Highlighting Collapsing container functional visually
[Zooming Container] Zooming in on hoveredcards

    The problem is, the zoom is infinite until the card is hovered and it never returns to its original size. The problem is, that we don't know what the original size is.

[Zooming Container] The zoom works (partially)

    The zoom only works on the top container in the current setup and also, player 2's character is missing.

[Zooming Container]

    Fixed the problem of only one container working

[Zooming Container] Fixed the card overflow issue

    The offset has been fixed according to the follwoing equations:
    containerWidth = cardWidth + {(containerWidth - cardWidth) = w}
    w = {x * cardWidth = offset} * (childrenCount - 1)
    x = w/(cardWidth * (childrenCount - 1))
    x = (containterWidth - cardWidth)/(cardWidth * (childrenCount - 1))
    offset = cardWidth * (containerWidth - cardWidth)/(cardWidth * (childrenCount - 1))
    offset = (containerWidth - cardWidth) / (childrenCount - 1)

+++ include/UtilsLib/LengthUnits.h
@@ -16,12 +16,12 @@ enum class LengthUnits : int32_t
   px = 2
 };
 
-inline auto ConvertLengthUnit(LengthUnits from, LengthUnits to) -> int32_t {
+inline auto ConvertLengthUnit(LengthUnits from, LengthUnits to) -> float {
   if (from == to)
     return 1;
   const int32_t magnitude =
     std::powf(10, static_cast<int32_t>(from)) * std::powf(10, static_cast<int32_t>(to));
-  return to > from ? 1 / magnitude : magnitude;
+  return to > from ? 1.f / magnitude : magnitude;
 }
 
 } // namespace Utils
